Some upgrades to facets
