+++ sfm_pc/utils.py
@@ -358,3 +358,41 @@ def import_class(cl):
     classname = cl[d+1:len(cl)]
     m = __import__(cl[0:d], globals(), locals(), [classname])
     return getattr(m, classname)
+
+def format_facets(facet_dict):
+    '''
+    pysolr formats facets in a weird way. This helper function converts their
+    list-like data structure into a standard tuple.
+
+    Basic idea: convert counts from a list to a list of tuples, e.g.:
+
+    ['foo', 1, 'bar', 2] --> [('foo', 1), ('bar': 2)]
+    '''
+    facet_types = ['facet_queries', 'facet_fields', 'facet_ranges',
+                   'facet_heatmaps', 'facet_intervals']
+
+    out = {}
+
+    for ftype, facets in facet_dict.items():
+        updated_facets = {}
+        for facet, items in facets.items():
+            if isinstance(items, dict):
+                # Ranges have a slightly different format
+                item_list = items['counts']
+            else:
+                item_list = items
+
+            counts = []
+            for i, el in enumerate(item_list):
+                # The attribute name always comes first; use them as keys
+                if i % 2 == 0:
+                    count = item_list[i+1]
+                    counts.append((el, count))
+                else:
+                    # We already bunched this one, so skip it
+                    continue
+
+            updated_facets[facet] = counts
+        out[ftype] = updated_facets
+
+    return out
