+++ templates/partials/search_filter.html
@@ -1,11 +1,11 @@
 <div class="panel panel-filter">
     <div class="panel-heading" data-toggle="collapse" data-target="#filter-{{facet_name}}">
-        <p class="panel-title">
-            <a role="button">
+        <a role="button" class="filter-heading">
+            <p class="panel-title">
                 <strong>{{facet_label}}</strong>
                 <i class="fa fa-fw {% if facet_name in selected_facets %}fa-minus {% else %} fa-plus {% endif %} pull-right"></i>
-            </a>
-        </p>
+            </p>
+        </a>
     </div>
     <div class="panel-body {% if facet_name in selected_facets %}panel-show{% endif %} collapse" id="filter-{{facet_name}}">
         <ul class="search-facet-list">
